+++ ricaun.NUnit.Tests/NUnitTests.cs
@@ -303,7 +303,8 @@ namespace ricaun.NUnit.Tests
         [TestCase("*.MultipleTest.*", 2)]
         [TestCase("*AbstractTest?*", 4)]
         [TestCase("*Abstract*", 4)]
-        [TestCase("*", 60)]
+        [TestCase("*.TestsCaseSource.*", 10)]
+        [TestCase("*", 70)]
         public void TestAssembly_Filter(string testName, int numberOfTests)
         {
             TestEngineFilter.Add(testName);